refactor: split paths.py into focused modules
Extract ConfigPaths dataclass and env file loader into separate internal
modules. Original file preserved as re-export entry point with default_paths.
