+++ src/quodeq/config/paths.py
@@ -1,13 +1,16 @@
-"""Filesystem path resolution for quodeq configuration directories."""
+"""Filesystem path resolution for quodeq configuration directories.
+
+Re-exports from internal modules to preserve the public API.
+"""
 
 from __future__ import annotations
 
-import logging
-import os
-from dataclasses import dataclass
 from pathlib import Path
 
-_logger = logging.getLogger(__name__)
+from quodeq.config._config_paths import ConfigPaths  # noqa: F401
+from quodeq.config._env_loader import load_env_file  # noqa: F401
+
+__all__ = ["ConfigPaths", "load_env_file", "default_paths"]
 
 # Number of parent directory levels from this module to the project root
 # (config -> quodeq -> src).
@@ -17,93 +20,8 @@ _FALLBACK_PARENT_DEPTH = 3
 _DATA_DIR = Path(__file__).resolve().parent.parent / "data"
 
 
-@dataclass(frozen=True)
-class ConfigPaths:
-    """Immutable set of resolved paths to all configuration directories and files."""
-
-    root: Path
-    prompts_dir: Path | None = None
-    standards_dir: Path | None = None
-    env_file: Path | None = None
-    gitignore_file: Path | None = None
-
-    @property
-    def evaluators_dir(self) -> Path:
-        """Global directory for custom evaluator JSON files."""
-        return Path.home() / ".quodeq" / "evaluators"
-
-    @property
-    def vroot(self) -> Path:
-        """Return the versioned root path (currently identical to root)."""
-        return self.root
-
-    @property
-    def disciplines_conf(self) -> Path:
-        """Return the path to the disciplines configuration file."""
-        return self.root / "config" / "disciplines.conf"
-
-    @property
-    def detection_file(self) -> Path:
-        """Return the path to the universal detection.json config."""
-        return self.root / "config" / "detection.json"
-
-    @property
-    def dimensions_file(self) -> Path:
-        """Return the path to the universal dimensions.json config."""
-        return self.root / "config" / "dimensions.json"
-
-    @classmethod
-    def from_root(cls, root: Path) -> "ConfigPaths":
-        """Construct a ConfigPaths instance by deriving all paths from a root directory."""
-        return cls(
-            root=root,
-            prompts_dir=root / "prompts",
-            standards_dir=root / "standards",
-            env_file=root / ".quodeq.env",
-            gitignore_file=root / ".gitignore",
-        )
-
-    @classmethod
-    def from_data_dir(cls, data_dir: Path) -> "ConfigPaths":
-        """Construct a ConfigPaths from the bundled package data directory."""
-        return cls.from_root(data_dir)
-
-
-def load_env_file(paths: ConfigPaths, target: dict[str, str] | None = None) -> None:
-    """Source all ``export VAR=value`` lines from ``.quodeq.env`` into *target*.
-
-    *target* defaults to ``os.environ``.  Already-set keys are NOT
-    overwritten, so explicit env takes precedence over the file.
-
-    Pass an explicit *target* dict in tests to avoid mutating the real
-    environment.
-    """
-    if target is None:
-        target = os.environ
-    if not paths.env_file.exists():
-        return
-    try:
-        lines = paths.env_file.read_text().splitlines()
-    except OSError as exc:
-        _logger.warning("Failed to read env file %s: %s", paths.env_file, exc)
-        return
-    for raw_line in lines:
-        stripped = raw_line.strip()
-        if not stripped or stripped.startswith("#"):
-            continue
-        if stripped.startswith("export "):
-            stripped = stripped[len("export "):]
-        if "=" not in stripped:
-            continue
-        key, _, value = stripped.partition("=")
-        key = key.strip()
-        value = value.strip()
-        if key and key not in target:
-            target[key] = os.path.expanduser(value)
-
-
 def _looks_like_project_root(root: Path) -> bool:
-    """Return True if *root* contains the expected quodeq data layout (prompts/ + config/detection.json)."""
+    """Return True if *root* contains the expected quodeq data layout."""
     has_prompts = (root / "prompts").is_dir()
     has_detection = (root / "config" / "detection.json").is_file()
     return has_prompts and has_detection
@@ -121,7 +39,7 @@ def default_paths() -> ConfigPaths:
         if _looks_like_project_root(root):
             return ConfigPaths.from_root(root)
 
-    # Walk up from src/quodeq/config/paths.py to project root (3 levels: config -> quodeq -> src)
+    # Walk up from src/quodeq/config/paths.py to project root
     if len(module_path.parents) > _FALLBACK_PARENT_DEPTH:
         root = module_path.parents[_FALLBACK_PARENT_DEPTH]
     else:
